requirements.txt
django==4.2
djangorestframework
djangorestframework-simplejwt
environs
psycopg2-binary
Pillow